Checkout the frontend submodule when not existing
Checkout the submodule during the build if it doesn't exist.
This could happen when a develop defaults to normal `git clone`
checkout behavior.

Do not pull the latest changes. Pulling puts the repo in a dirty state
if the frontend code has new commits. That may not be desired if what
you are after is the latest known safe build.