Fix /api/v1/auth/clients test to check all guids
Before the test, we create a OAuth2Client object and then we invoke
/api/v1/auth/clients and expect the OAuth2Client object to be returned
as the first item in the list.  This appears to always work in sqlite
but in postgresql, the first item is usually not the one we created in
the test fixture causing the test to fail.

I changed the assertion in the test to look through all the OAuth2Client
objects returned by the api.