Refactor build frontend
Corrects a ownership issue due to this script now building in a docker
container. The artifact results of the build no longer place
themselves in the root of the project, thus are no longer pose a
removal problem to the user. Unpacking and installation of the built
distribution is now done on the host machine rather than in the docker container.