Only upload coverage to codecov for ubuntu py3.9
Coverage changes as we run the tests for macos, ubuntu sqlite,
postgresql and different versions of python, so you only get the final
result for coverage after all the tests are done.  I would like to take
a shortcut and only run coverage for one version of python and since
ubuntu and python 3.9 is what we are using for production, I'm picking
this as the one to use.