Add docker examples for wildme/houston in README
There was an example in the README for how to run
`frolvlad/flask-restplus-server-example` which is completely unrelated
to houston.  Change it to use the local `Dockerfile` and specifically
how to run the tests.

The entrypoint for the wildme/houston file at the time of writing
ignores `-u root` in docker run to run as root and causes permission
problems when using `pip` etc so disabling the entrypoint by doing
`--entrypoint=`.