Add script to list all endpoints in matching order
This is just a command to list all the endpoints registered in the
application and the order they are matched within the appliation.

This gives us a comprehensive list of endpoints. This can also help in
debugging around url to view matching. For example, it appears that
any url registered with the GET method automatically gets the OPTION
method registered.