Integrate frontend development within the app suite
Change that attempts to bring in a frontend developer story. The idea
is that the (frontend|backend) developer can develop with all the
components available to them.

I've used a `dev-` prefix on the frontend component within the
component composition to indicate this component is only for
development usage. It's probably not obvious and doesn't really need
to be, but just wanted to mention incase that came into question.