Enable reverse proxying in compose setup
This more or less makes both nginx and houston aware that they are in
a proxy situation.

Setting `Host` to `$http_host` in nginx allows proxied 302s to
successfully rewrite the Location header to an absolute url. nginx
does this rewrite by default and uses the `Host` value in the
operation. There is another way to do this, but this was easiest and
more dynamic, which is what you want for a development scenario anyhow.