Add CI steps to publish the latest image
Changes the existing publishing step, which only works on tag right
now. This no longer pushes the `latest` as the latest release. Instead
this publishes the tip of `develop` as the `wildme/houston:latest` image.

I wanted to be able to release the `latest` tagged image, so we aren't
pulling an out of date image when a tag is not specified. We're not
currently doing tagged releases (yet?). Maybe when we do, we can
change this behavior back and create a `develop` image tag.