Change tus test using session id to use any session directories
We have been getting test failures where the test tries to look for
`_db/uploads/session-<hashed-session-id>/a.txt` and failed because
`_db/uploads/session-<hashed-session-id>/` doesn't exist.  It happens
very rarely for me locally so it was really hard to debug.

I'm now changing the test to look in any session directories so
`_db/uploads/session-*/a.txt` and assert that there's at least one.
:crossed_fingers: