Make the houston .env configuration load env-var definable
We require the .env file that is loaded by Houston to remain between
container cycles. The data that remains between cycles is going to be
within the data volume. This change puts the .env within the data
volume, defines it's location as $HOUSTON_DOTENV, and allows the
user to define that location while maintaining a default.