Update the project's dependencies at a high level
I've removed the python dependencies from the list, because that's
minutia that isn't directly user applicable. That is, once the user
has Python installed, we take care of the minutia of installing the
application's dependencies. What we don't take care of are installing
the high level components for the user (unless they use the docker
setup).