Remove docs for build and patched dependencies
These aren't particularly useful. We've mostly had to work around the
build dependencies bit of this anyway. So yes, pyinvoke is kinda
required, but it'd be more work to explain why this is the case than
to just have the user install the application via pip and then have
this requirement available.

The patched dependencies docs explain the upstream template's decision
process, but are less applicable to the long-term needs of this project.