Create expected gitlab repos as test fixtures
The tests expect there to be 2 gitlab repos
`00000000-0000-0000-0000-000000000002` and
`00000000-0000-0000-0000-000000000001`.  These repos are created using
`invoke app.initialize.initialize-gitlab-submissions --email test@localhost`.

Since we are starting with an empty gitlab instance in docker-compose,
it would be better if these repos are created as part of the test
fixtures.

Now the code for creating the repos are in `test_submission_uuid` and
`test_empty_submission_uuid`.