Add cropping to user profile image API
To crop a user profile image:

```
PATCH /api/v1/users/<user-guid>

{"op": "replace",
 "path": "/profile_fileload_guid",
 "value": {
   "guid": "<fileupload-guid>",
   "crop": {"x": 650, "y": 150, "width": 150, "height": 150}
 }
}
```

The `op` can be "replace" or "add".

The `value` requires either "guid" or "transactionId".

This modifies the image and store the cropped image only.  The reason is
specifically because this is a user profile image, if a user uploads a
family picture, it's probably best to not store the original.