Move non-essential testing to nightly schedule
This moves the non-essential testing paths to the nightly scheduled
workflow. The goal is to test the critical path, that which we deploy
with, for change acceptance.
In the event that the changes do cause problems on the mac
or in other versions of python we'll be notified the day after.
But these problems shouldn't block deployable progress.

We will still be testing both SQLite and Postgres in the PR tests. But
this change only runs those tests on 3.9, our deploy version. We're
testing both database implementations, because database integration
can sometimes be tricky to get right; and testing both is rather
inexpensive (in time).