Add PERMISSION_ROLE_MAP in permissions rules
The current approach to permission is very hard to understand because
there is a lot of custom code to figure out if a user has access to
something.  There are multiple layers of checking.  It would be much
easier to understand if we just use:

```
SiteSetting, READ, Module => is_privileged, is_admin
```

A declarative approach means that we can quickly see what permissions
different roles have and which roles have a particular permission.