Set request content type to application/json in tests
In preparation for the upgrade to webargs 5.5.3, which does not try to
parse the request data unless the content type of the request is
application/json:

CVE-2020-7965
Published 2020-01-29T15:15:00
flaskparser.py in Webargs 5.x through 5.5.2 doesn't check that the
Content-Type header is application/json when receiving JSON input. If
the request body is valid JSON, it will accept it even if the content
type is application/x-www-form-urlencoded. This allows for JSON POST
requests to be made across domains, leading to CSRF.