fix project lookup to find a singular project
In the tests the 000...002 type uuid projects exist. If the gitlab
service has more than one of these projects, the incorrect project can
be found. This most recently happened because a second namespace/group
was created named `TEST-from-fig`, which also matches the intended
namespace `TEST`. Both groups/namespaces have this project.

This change ensures that the search for the project is within
specified group/namespace.