make specific blueprint for swagger-ui customizations
This enables us to set exactly where the swagger-ui dist is
located. It was previously assumed to be located within the static
directory, but in development this may not be the case or the
intention. So this change gives the developer space to locate the
built swagger-ui files wherever they'd like.

The goal is to allow the container to build the ui in the image, but
also allow the developer to mount over top of it in development.