Implement local alternative to gitlab
If GITLAB_REMOTE_URI is set to a file path, the local alternative to
gitlab is used, for example:

```
GITLAB_REMOTE_URI=/code/_db/git_remote/ pytest
```

runs the tests using git repositories locally without gitlab.

This is currently only for development.  This was done to try to allow
local development without a gitlab instance which uses up a lot of
memory.