DEX-282 Create dedicated test database and data directories
When we ran `pytest` in the houston container, it wiped all the tables
from the database when it was done.  This is the correct behavior for
tests but if we are also running the site with the same database, we
would have to reset the database before we can run the site again.

To fix this problem, we are creating a dedicated test database (by
specifying `TEST_DATABASE_URI`) and data directories (using
`tempfile.TemporaryDirectory`).  So it is now possible to run the site
and tests without them interfering with each other.