Fix GITLAB_REMOTE_LOGIN_PAT not loaded in app config
When using docker-compose, in the houston entrypoint
`.dockerfiles/docker-entrypoint.sh`, we try to load
`GITLAB_REMOTE_LOGIN_PAT`:

```
file_env 'GITLAB_REMOTE_LOGIN_PAT'
```

Since there is nothing set up, we end up with `GITLAB_REMOTE_LOGIN_PAT`
set to an empty string.

In `config.py`:

```
load_dotenv(_dotenv, override=False)
```

Which is supposed to load `GITLAB_REMOTE_LOGIN_PAT` in `/data/var/.env`,
but since it is already in `os.environ` and `override=False`,
`GITLAB_REMOTE_LOGIN_PAT` ends up not getting loaded.

The fix is in the entrypoint, to not set `GITLAB_REMOTE_LOGIN_PAT` if it
is an empty string.