Fix create sighting multiple assets not stored
When creating a sighting with multiple assets, `_validate_assets` failed
because:

```
if len(assets) != len(paths_wanted_list):
```

`assets` was `[<Asset ...>, <Asset ...>]` and `paths_wanted_list` was
`[{'zebra.jpg', 'fluke.jpg'}]` which means `len(assets)` was 2 and
`len(paths_wanted_list)` was 1 causing validation to fail.

In `_validate_asset_references`, the code returned a list of sets
because it was originally written for assets per encounter.  Now that
the assets are in the sighting, we don't need the list anymore.

Update `test_create_anon_and_delete_sighting` to include 2 assets and
check the output for assets.

When deleting the sighting at the end of the test, there was an
`AssetGroup` left over.  It appears in `Asset.delete_cascade`, we need
to do `db.session.refresh(asset_group)` after deleting the asset so the
asset group doesn't see the asset anymore.