Add check to all tests for clean up of AssetGroups, Sightings etc
Add an autouse fixture "check_cleanup_objects" which is applied to all
the tests to make sure the number of asset groups, assets, sightings,
encounters stay the same after the tests.

Use `mock.patch('app.modules.asset_groups.tasks.delete_remote')` to stop
`asset_group.delete()` from deleting the gitlab project.  This is for
the asset group fixtures shared between tests.  Gitlab projects are not
deleted immediately and sometimes we get to a state where gitlab is in
the middle of deleting the project while we're trying to create it
again.  So let's just avoid deleting it.