Store AssetGroupSighting.jobs correctly in the database
The first problem is in the jobs column definition, we had `default={}`
and then we modify `self.jobs` by doing:

```
self.jobs[str(job_id)] = {...}
```

The default `{}` is only initialized once, so we are basically editing
the same dict in different AssetGroupSighting instances.

We fix this by doing `default=lambda: {}` so we are creating a new
instance of dict for every AssetGroupSighting.

Next, the AssetGroupSighting isn't saved after we modified `self.jobs`.
The problem here is modifying `self.jobs[str(job_id)]` doesn't trigger
the sqlalchemy code to detect something has actually changed.
`self.jobs` needs to be assigned to in order for this mechanism to work:

```
self.jobs = self.jobs
```

Finally, sqlalchemy would compare the old value of `self.jobs` with the
new value of `self.jobs` before the assignment and of course `self.jobs`
was modified, so the comparison would come back "the same".  So the fix
here is done in the `JSON` type definition in
`app/extensions/__init__.py`:

```
def compare_values(self, x, y):
    return False
```

This tells sqlalchemy that the values are not the same and so it should
write to the database.