Store Sighting.jobs correctly in the database
The first problem is in the jobs column definition, we had `default={}`
and then we modify `self.jobs` by doing:

```
self.jobs[str(job_id)] = {...}
```

The default `{}` is only initialized once, so we are basically editing
the same dict in different Sighting instances.

We fix this by doing `default=lambda: {}` so we are creating a new
instance of dict for every Sighting.

Next, the Sighting isn't saved after we modified `self.jobs`.
The problem here is modifying `self.jobs[str(job_id)]` doesn't trigger
the sqlalchemy code to detect something has actually changed.
`self.jobs` needs to be assigned to in order for this mechanism to work:

```
self.jobs = self.jobs
```