Change backref to back_populates in db.relationship
Using `back_populates` instead of `backref` means we have to define the
relationship in both classes instead of auto generating it in the other
class.  It might be easier to understand to the reader when looking at
the code and it will allow us to specify "order_by" when it is a
one-to-many relationship.

Import all the models first in `app.modules.init_app` before doing other
stuff to avoid model lookup error.  Do something similar in
`tests/conftest.py` so importing DetailedUserSchema works (
`pytest tests/test_transaction.py`)