Create oauth user for acm before first request
Add code to create oauth user for acm to use before first request.  ACM
needs a oauth client id and secret in order to get the images from
houston.  This adds a static client id and secret which is shared by
houston and acm and it is created before the first request so ACM will
be able to authenticate.