Fix minor bugs in AssetGroupSighting get_completion and complete
When calling `AssetGroupSighting.get_completion` when the stage is
"detection":

```
app/modules/asset_groups/models.py:271: in get_completion
    complete_jobs = [job for job in self.jobs if not job['active']]
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

.0 = <dict_keyiterator object at 0x7f9e849aa310>

>   complete_jobs = [job for job in self.jobs if not job['active']]
E   TypeError: string indices must be integers

app/modules/asset_groups/models.py:271: TypeError
```

This is because `self.jobs` is a dict which means `for job in self.jobs`
is the `job` is actually the job id.  Changing `self.jobs` to
`self.jobs.values()` fixes the problem.

When calling `AssetGroupSighting.complete` to change from "curation" to
"processed", it checks all the jobs are active but all the jobs are set
to inactive after `AssetGroupSighting.job_complete` to change from
"detection" to "curation", this is probably a bug so change it to check
that all the jobs are inactive.