Move derived file ext and mime type to constants in Asset
We hardcoded the content type to be "image/jpeg" in resources.py and
the file extension to be "jpg" in models.py.  Probably best to move them
to constants so it's easier to see what's going on.

By the way, I tried sending a png file to sage detection and it came
back with results so I guess it's working even when it's not jpegs.