Remove app.env.enter task in favor of app.shell
I'm moving to this approach, because you don't want the environment to change from under you, which this task does with installation and upgrade. Quite often the shell is used to proof a problem and if in the process of proving it things change, the problem just gets harder to debug. So I'm removing that implicity logic in favor of the user explicitly installing or upgrading separate from running the interpreter/shell.