Replace healthcheck for starting celery containers with wait-for
We only use healthcheck for knowing when to start the celery containers
and is not used afterwards.

The problem with healthcheck is it creates lots of extra useless logs
and also causes `docker-compose up -d` to fail regularly in CI when the
containers are unhealthy.

We can use "wait-for" to wait until houston web server is up instead.

Also change the houston container to wait for elastic search to be up.
We connect to elastic search during initialization so not being able to
connect causes the houston container to exit.