Change social-groups PATCH to only use /members
Remove `/member` PATCH path and instead use `/members` to be consistent.
When op is "add", the members are added (or replaced if they already
existed).  When the op is "removed", the member or list of members are
removed.  When the op is "replaced", the members are set to exactly what
the user provides.