Fix collaborations created by user managers
If a user manager creates a collaboration between user1 and user2 and
the user manager already has a collaboration with user1, houston
responds with
"User <user-manager> attempted repeated collaboration with <user1>" and
returns the collaboration between user manager and user1.

The reason is because the code first checks whether there's a
collaboration between logged in user and user1, even if the
collaboration is for user1 and user2.