Fix gitlab-cleanup.yml houston timeout waiting for gitlab
The gitlab-cleanup action is timing out on CI because houston is waiting
for the gitlab container to be up.

We are not including gitlab in `docker-compose up` because in
gitlab-cleanup.yml, we are trying to clean up repos that are in
https://sub.dyn.wildme.io/.  We need to change `.env` so houston knows
which gitlab server we are using.

Also change `set_up_gitlab.sh` to skip if `GITLAB_REMOTE_LOGIN_PAT` is
already set.