Add integration test for asset group sightings
Also fix asset group sighting commit API sometimes getting errors from
EDM when the user's request does not have the
`Content-Type: application/json` header:

```
2021-11-08 11:29:31,501 [INFO] [app.extensions.restManager.RestManager] Sending post request to EDM: http://edm:8080/api/v0/org.ecocean.Occurrence/
2021-11-08 11:29:31,501 [DEBUG] [app.extensions.restManager.RestManager] Contents {'data': {'startTime': '2000-01-01T01:01:01Z', 'locationId': 'PYTEST', 'encounters': [{'guid': 'd040bd6e-8187-4a6f-ae52-bec9f491b558'}], 'assetReferences': ['zebra.jpg']}, 'headers': {'Accept': '*/*', 'User-Agent': 'python-requests/2.26.0'}}
2021-11-08 11:29:31,517 [AUDIT] [app.extensions.edm] Failed: sighting.data post failed {'args': {'class': 'org.ecocean.Occurrence'}, 'details': 'org.ecocean.api.ApiValueException: startTime must have a value', 'key': 'error'} 601 400 executed by user :146520af-fde5-41de-8be7-096c936c2a1d root@example.org
2021-11-08 11:29:31,519 [ERROR] [app.modules.utils] Bailing on AssetGroup creation: Sighting.post failed (error_fields ['startTime'])
```

The problem is with "Contents {'data': {'startTime': ...", it should be
"Contents {'json': {'startTime': ...".  Because the original request
doesn't have "Content-Type: application/json", RestManager does not send
it as json.

Instead of relying on the request having the json content type header,
just add it to the code that's calling the EDM API.