Allow configuraiton of the full filepath to the SSH key
In the orchestration scenario the ssh key is mounted on the filesystem
from a [kubernetes] secret. This location needs to reside outside of
the mounted data volume, because a mount can't contain a mount.

This change allows the current practice to continue working when
`GIT_SSH_KEY_FILEPATH` is undefined.

A problem would arrise if both `GIT_SSH_KEY` and
`GIT_SSH_KEY_FILEPATH` are supplied. But if the
permissions are set correctly to read-only on the file at
`GIT_SSH_KEY_FILEPATH`, the operation should fail with an OSError.