Organize the docker ignores
The hope is that this will help us understand why an ignore was put in
place. Generally speaking we don't want personal exceptions to be in
the ignore list, because that would create a huge file. Though,
personal files shouldn't really be an issue, because no one should be
publishing the container from their local machine (CI does that for us).