Add a development build target to the container instructions
The logic to resolve developer dependencies (e.g. testing
dependencies) has been removed from the entrypoint script. This logic
is now moved into the `development` build target.

This change attempts to mitigate connection issues by ensuring the
developer can continue working in the absense of a good connection to
pypi. That is to say that the built container will no longer request
packages on first run or restart, but instead will have them built in.

This will however require the developer to re-build the image when
dependencies are added to the code, as opposed to them magically
getting resolved on a restart of the containerized service.