DRY the container and app invocation
Changes the container build to a generic usage. There are no
differences at this time between how the variants houston startup.

If adjustments need to be made to the startup code for one of the
variants, please make the attempt to make such changes in
configuration rather than in the code.

The `PROJECT` build argument can't yet be removed because it's still
in use by the container for pulling and building the frontend.