Fix sage detection with multiple assets
When multiple image urls were sent to sage, sage was only using the
first url causing houston to reject the callback because the number of
assets didn't match.

It turns out the format we sent the image urls in was wrong...  we sent
multiple "image_uuid_list=" (normal form post) but sage expected a json
encoded list of strings.

Also add an integration test that exposed this bug. (data provided by
Ben)

Also retry sage_detection task when there's a sqlalchemy error because
it was happening a lot locally and the jobs field was "{}".