Remove the gitlab cleanup workflow
This was intended to routinely cleanup the gitlab instance. This was
originally done in github because it seemed like a good place to do it
at the time. Now with the gitlab service down most of the time, it
makes sense to just remove this workflow and bring it back when and if
needed.