Remove action on import
Let the application code be the defining entity for creation and
management of the various storage directories. In other words, don't
do this operation without access to the application's fully
instantiated configuration.

This goes along with the principals of functional programming, just at
the application level rather than the function level. Essentially,
don't produce side-effects as a part of your desired procedure.