Config package to allow for multiple contexts and environments
Allow the contexts and environments to live together in harmony.

This is a breaking change. I've choosen to break this change into
several commits, this being the foundation of those changes.

The idea here is to allow the developer to acquire the configuration
separate from the application, because that's how the code is
currently used in some places.

This introduces two method: `get_preliminary_config` and
`configure_app`. The goal here is to be able to get the configuration
without the application using `get_preliminary_config`. And then the
`configure_app` to associate configuration with the flask application.