Move dev tools requirements from houston Dockerfile to app/requirements.dev.txt
There is a version conflict for click in the packages we install:

```
ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
flask 1.1.4 requires click<8.0,>=5.1, but you have click 8.0.3 which is incompatible.
celery 5.1.2 requires click<8.0,>=7.0, but you have click 8.0.3 which is incompatible.
```

click 8.0.3 was installed because of ipython which requires black.  The
latest version of black 22.1.0 requires `click>=8.0.0` which is
incompatible with our flask and celery versions.

This could be solved by downgrading black but black only has one version
on pip so the way to solve this is to install an older tag from their
github repo.  I feel like this is better in a separate requirements file
since the github url is kind of long.