Use SERVER_NAME and PREFERRED_URL_SCHEME in the composition
This attempts to use the `SERVER_NAME` and `PREFERRED_URL_SCHEME`
environment variables. This is a move towards deprecating the
`HOUSTON_URL` environment variable that sets the `BASE_URL`
configuration setting. See also
[DEX-695 Obsolete the BASE_URL configuration
setting](https://wildme.atlassian.net/browse/DEX-695)

One downside of this change is that `SERVER_NAME` setting has special
behavior within Werkzeug. The server, much like nginx or any other
server, will only serve for the given name; so the `Host:
$SERVER_NAME` request header must be supplied when making requests
within the houston container if trying to directly contact the service
without the reverse proxy.

On the upside, because we tell Werkzeug to serve under this name,
we must also be able to actual server under the name and share it with
other services (i.e. ACM in the callback url). To do this I've worked
around the issue by utilizing the name `localhost` for both internal
(to the composition's network) and external usage. For example,
requesting `http://localhost/api/v1/users/me` will work both inside the
container and outside of it.

Using this change will mean the composition is serving very similar to
how it would be served in production (or at least how Flask likes to
be served in production). And thus no workaround, other than the
service named `localhost` added in this change, will need to be
present. That means cleaner code with less special snowflakes! Yay!