Dispose sqlalchemy engine after elasticsearch load_codex_indexes task
The elasticsearch load_codex_indexes task is run periodically in the
celery_worker container.  It seems the wildbook and houston sqlalchemy
engines hold all the connections open until the variables go out of
scope, but it never happened for some reason, so we can force it by
doing `engine.dispose()`.