Mount all of Houston into frontend container
In order to get the git information from the frontend code (e.g. run
`git` commands in the `_frontend.codex` directory) the parent project
must be accessible. This is only a problem because the `_frontend.*`
repos are submodules; and thus not actually git repos
themselves (i.e. it's complicated). This solves `fatal: No names
found, cannot describe anything.` when running git commands within the
submodule.

This change is being made, because additions are being made to the
frontend that require git info. The frontend build code discovers its
own version information for display in the server-status page. This
fix is required to enable that feature.