Fix dev-frontend entrypoint to assume current working directory
This will make the assumption that the current working directory (CWD)
is the root of the frontend code. But if the submodule directory name is
present in the CWD, change to it before running entrypoint commands.

This enables the rewriting of the volume mount from the docker-compose
file without changing the `working_dir` setting as well.