Remove HoustonConfig implemenation in favor of environment variables
It's important to distinguish between Houston Config and Site Settings. General Configuration, which the Houston Config and Flask Config are a part of, are application settings that need to be configured before app startup, configuration that is read-only and shouldn't be changed by the frontend users (e.g. database connection string), and require restart to take effect. Site Settings on the other hand are settings the site admins could change and expect the site to utilize without restart (e.g. site name, email address). So two very different things that most notably has to do with end-user mutability and service restart.

AFAICT we are not using Houston Config. I say this with some degree of certainty, because I know I'm not using it in the deployment code. And last I recall, it's not being used in the docker-compose setup.

The primary means of configuring the application is through *environment variables*. In development we do this through docker-compose, which happens to use the `.env` file for ease of use. In deployment on azure the settings come out of `etcd`, a key store database... Think of it like the `.env` file equivalent, but not on the filesystem. The use of environment variables is what gives us the ability to *deploy the application code for different instances without changing the code*.

I heard the use case for Houston Config is to allow for the deployment of multiple sites using the same codebase. I'd say that is already addressed and does not use Houston Config to achieve that end. The confusion seems to be around the flat `.env` configuration file as something that is needed for every deployment. This is not the case. The `.env` file is for ease of use in defining environment variables. The `.env` provided in the repo is for developer use. (note, it is ignored by docker; see `.dockerignore`).

So what I'm trying to say, is that we've got this HoustonConfig implementation that requires the database before application instantiation and startup. This change removes it because we don't use it and the use case it was meant to solve is addressed in a different and probably more robust way.