Don't use request's User-Agent in RestManager requests
ACM has a module called `flask_paranoid` which creates a session token
using the request's User-Agent and IP address.  It compares this session
token during every request.  If the User-Agent changes, causing the
session token to be invalid, `flask_paranoid` clears the session and
redirects the user to `/`.

We found this error when accessing acm's `assets.list` API.  Sometimes
it would return json, other time it would redirect the user to `/`.