Add USE_RELOADER environment variable for flask reloading
Instead of reloading in the development environment, use `USE_RELOADER`
environment variable to turn on and off automatic reloading.

The reason for this change is because if a file is saved with syntax
errors, flask reloads and fails with an exception, the houston container
dies and this kills all the shell in the container and so not everyone
wants that behavior.