Rerun detection in foreground in celery, fixup for 0e315a0c
We have periodic jobs that check if detection or identification should
rerun, since the checking is already done in celery we can just call the
sage detection or identification API in the foreground instead of
queueing them.

This is to bypass a problem on codex-qa I see right now where queued
jobs are not being picked up but periodic jobs are still working in
celery.