Fix AssetGroupSighting detection and curation start time
Frontend is using detection and curation to determine what state the
object is in.  If detection start time is none and curation start time
is not none, detection is shown as "skipped".

That was exactly the situation we were in on codex-qa because
detection start time was only returned if `self.jobs` was populated, and
`self.jobs` wasn't populated for a few minutes because of celery
problems.  In the mean time, curation start time was returned because
there were no active jobs.

Change the code to check `self.stage` in addition to the original logic.