Add environment variables for default email settings
This is so we can set email settings without touching the database for
our own sites.  This doesn't affect the ability to set email settings in
the database and from the frontend.  The environment variables are only
defaults.

Now that the default may contain sensitive data, let's just not return
any non public fields to non admin users when they fetch all the
configuration.