Fix initial and error values for elasticsearch test check
This is strangely coded. I don't quite follow how this works. Afaict
it's never broken for anyone else.

The response from get_elasticsearch_status is a dict. So the list with
a single None values doesn't make a whole lot of sense. Checking the
lenght of a dictionary is an odd thing to do, but it works. This is
what the later condition in this function does to check whether it
should break from the loop.

Sticking with the intention, allow the function to loop back around
after an error. This seems like a good thing.