Removes gumby managed indexes
This change removes the gumby managed indexes that are no longer used
by the frontend.

The primary focus of this changeset has been to remove the unnecessary
indexing of individuals, encounters and sightings. The secondary, but
important to other work, focus is to remove the need for the
`postgres_fdw` extension within the EDM database. This removal of the
indexing code will remove the need for those foreign table
definitions.

This does not remove gumby itself. That will come in a later changeset