Use the default config variable when env-var is empty
The problem with `os.getenv(key, default_value)` is the `default_value`
is only returned when the environment variable isn't set.  Sometimes we
want to return the `default_value` when the environment variable is an
empty string.

Add `config.utils._getenv` which returns the default if the environment
variable is an empty string.  But if `empty_ok=True` then it returns an
empty string if the environment variable is an empty string.

Convert all `os.getenv` and some of the `os.environ` to
`config.utils._getenv`.