DRY the sentry config and use a single variable name
I feel centralizing the setting makes sense; thus DRY the code.
I can see the desire to distinguish between development vs production,
especially when you have a default for development, but I'm going to
opt for non-default telemetry. Development errors are going to
be many; so providing a default means lots of noise in sentry that
will result in no one using sentry.