Always run data directories creation code in docker-entrypoint.sh
The original problem was log files directory was created only if the
houston volume was not initialized, which means people have to clear
their houston volume for the log files directory to be created.  I've
changed it so the log files directory creation code is run whether the
houston volume is initialized or not.

The code includes changing the owner of the data directories to nobody
so the app can write to it.  The code used to be inside the "not already
initialized" block, but really it can run every time.