DEX-1174 Return sighting time if encounter time is missing in BaseEncounterSchema
When the encounter time is null, we are supposed to return the sighting
time. But that wasn't the case in the individual API.

The individual API uses `AugmentedEdmEncounterSchema` which uses
`Encounter.get_time_isoformat_in_timezone` and
`Encounter.get_time_specificity` for the time and time specificity.
Those methods should use `Encounter.get_time` with
`sighting_fallback=True` in order to fix this.

The change from:

```
time = base_fields.Function(Encounter.get_time_isoformat_in_timezone)
```

to

```
time = base_fields.Function(lambda enc: enc.get_time_isoformat_in_timezone())
```

is necessary because in the first case, `sighting_fallback` is `{}` for
some reason, in the second case, `sighting_fallback` is `True` as
expected.